refactor: separate music page and songs list
