+++ Screenbox/Pages/MusicPage.xaml.cs
@@ -1,9 +1,8 @@
 ﻿#nullable enable
 
-using Windows.UI.Xaml;
+using System;
 using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Controls.Primitives;
-using Windows.UI.Xaml.Input;
+using Windows.UI.Xaml.Media.Animation;
 using Windows.UI.Xaml.Navigation;
 using Microsoft.Extensions.DependencyInjection;
 using Screenbox.ViewModels;
@@ -15,8 +14,11 @@ namespace Screenbox.Pages
     /// <summary>
     /// An empty page that can be used on its own or navigated to within a Frame.
     /// </summary>
-    public sealed partial class MusicPage : Page
+    public sealed partial class MusicPage : Page, IContentFrame
     {
+        public object? FrameContent => ContentFrame;
+        public Type ContentSourcePageType => ContentFrame.SourcePageType;
+        public bool CanGoBack => ContentFrame.CanGoBack;
         internal MusicPageViewModel ViewModel => (MusicPageViewModel)DataContext;
 
         public MusicPage()
@@ -25,41 +27,20 @@ namespace Screenbox.Pages
             DataContext = App.Services.GetRequiredService<MusicPageViewModel>();
         }
 
-        protected override async void OnNavigatedTo(NavigationEventArgs e)
+        protected override void OnNavigatedTo(NavigationEventArgs e)
         {
-            VisualStateManager.GoToState(this, "Fetching", true);
-            await ViewModel.FetchSongsAsync();
-            VisualStateManager.GoToState(this, ViewModel.GroupedSongs?.Count > 0 ? "Normal" : "NoContent", true);
+            base.OnNavigatedTo(e);
+            ContentFrame.Navigate(typeof(SongsPage), null, new SuppressNavigationTransitionInfo());
         }
 
-        private void SongListView_OnRightTapped(object sender, RightTappedRoutedEventArgs e)
+        public void GoBack()
         {
-            if (ItemFlyout.Items == null) return;
-            if (e.OriginalSource is FrameworkElement { DataContext: MediaViewModel media } element)
-            {
-                foreach (MenuFlyoutItemBase itemBase in ItemFlyout.Items)
-                {
-                    itemBase.DataContext = media;
-                }
-
-                ItemFlyout.ShowAt(element, e.GetPosition(element));
-                e.Handled = true;
-            }
+            ContentFrame.GoBack();
         }
 
-        private void SongListView_OnContextRequested(UIElement sender, ContextRequestedEventArgs args)
+        public void NavigateContent(Type pageType, object? parameter)
         {
-            if (ItemFlyout.Items == null) return;
-            if (args.OriginalSource is SelectorItem item)
-            {
-                foreach (MenuFlyoutItemBase itemBase in ItemFlyout.Items)
-                {
-                    itemBase.DataContext = item.Content;
-                }
-
-                ItemFlyout.ShowAt(item);
-                args.Handled = true;
-            }
+            ContentFrame.Navigate(pageType, parameter);
         }
     }
 }
